Test ConfigProvider
