+++ package.json
@@ -53,6 +53,7 @@
     "@babel/core": "^7.12.16",
     "@babel/runtime": "^7.12.13",
     "@react-native-community/eslint-config": "^2.0.0",
+    "@testing-library/react-native": "^8.0.0",
     "@types/jest": "^26.0.20",
     "@types/react-native": "^0.63.50",
     "@types/react-native-vector-icons": "^6.4.6",
@@ -91,6 +92,7 @@
     "testPathIgnorePatterns": [
       "/node_modules/",
       "/e2e/"
-    ]
+    ],
+    "testRegex": ".*\\.test\\.ts.*$"
   }
 }
